fix: uses num type instead of double in ticker model and entity

+++ lib/features/landing/presentation/widgets/place_order_widget.dart
@@ -465,7 +465,7 @@ class _PlaceOrderWidgetState extends State<PlaceOrderWidget> {
             WidgetsBinding.instance
                 ?.addPostFrameCallback((_) => _onPriceAmountChanged(
                       placeOrderCubit,
-                      newPrice,
+                      newPrice.toDouble(),
                       false,
                     ));
           }
